refactor: 유저는 하나의 이미지만 가진다. (기획 변경)
- Account에 tag 연관관계 추가
- AccountImage, AccountTag에서 Account 연관관계 제거
- 여러장 이미지 등록하는 로직 변경
